+++ tests/test_packageurl.py
@@ -386,3 +386,17 @@ def test_vers_validation_ok():
 def test_vers_validation_fails():
     with pytest.raises(ValueError, match="must start with the 'vers:' URI scheme."):
         PackageURL.from_string("pkg:pypi/requests?vers=>=2.0")
+
+
+def test_no_encoding_to_string():
+    p = PackageURL(
+        type="nuget",
+        namespace="an:odd:space",
+        name="libiconv: character set conversion library",
+        version="1.9",
+        qualifiers={"package-id": "e11a609df352e292"},
+    )
+    assert (
+        p.to_string(encode=False)
+        == "pkg:nuget/an:odd:space/libiconv: character set conversion library@1.9?package-id=e11a609df352e292"
+    )